Refactor API
- create API dir
- move all relevant files to API dir
- rename Users dir to Schemas
- run server test (pass)
